+++ src/main/resources/db/migration/V3__create_permisstions_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE permissions
 (
-    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id          INT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(255) NOT NULL UNIQUE,
     description VARCHAR(255),
 
     created_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    deleted_at TIMESTAMP NULL
+    deleted_at  TIMESTAMP NULL
 );
